Update random number generator to return a double instead of a float (seems to have been done by the compiler already - but this makes it official)

+++ include/random_dirty.h
@@ -10,7 +10,7 @@ class random_dirty {
   random_dirty(long seed);
   
   // get new random value
-  float random_num();
+  double random_num();
 
  private:
   long _idum;
